Update echo server to handle object instead of []DataReading
